+++ openradar/publishing.py
@@ -156,7 +156,10 @@ class Publisher(object):
 
     def publish_image(self, cascade=False):
         """ Publish to geotiff image for webviewer. """
-        images.create_png_for_animated_gif(self.image_publications())
+        for publication in self.image_publications():
+            images.create_png_for_animated_gif(publication)
+            # also process the previous one to reduce the flickering
+            images.create_png_for_animated_gif(publication.previous())
 
     def publish_ftp(self, cascade=False, overwrite=True):
         """ Publish to FTP configured in config. """